requirements.txt
# requirements.txt
replicate
python-dotenv
pandas
numpy
matplotlib
seaborn
wordcloud
jupyter
